refactor(ballot-interpreter): collapse `timing_marks::corners` into `timing_marks`
There's only the one algorithm now, so we don't need this extra level of
hierarchy.
