+++ libs/ballot-interpreter/src/bubble-ballot-rust/timing_marks/mod.rs
@@ -7,12 +7,120 @@ use types_rs::{
     geometry::{GridUnit, PixelPosition, Point, Rect, Segment, Size, SubGridUnit, SubPixelUnit},
 };
 
-use crate::ballot_card::Geometry;
+use crate::ballot_card::{BallotImage, Geometry};
+use crate::interpret::Error;
 use crate::scoring::UnitIntervalScore;
 use crate::timing_marks::scoring::CandidateTimingMark;
 
-pub mod corners;
+pub mod border_finding;
+pub mod corner_finding;
+pub mod mark_finding;
 pub mod scoring;
+pub mod shape_finding;
+pub mod util;
+
+use crate::timing_marks::util::CornerWise;
+
+pub struct Options {
+    shape: shape_finding::Options,
+    corner: corner_finding::Options,
+    border: border_finding::Options,
+}
+
+impl DefaultForGeometry for Options {
+    fn default_for_geometry(geometry: &Geometry) -> Self {
+        Self {
+            shape: shape_finding::Options::default_for_geometry(geometry),
+            corner: corner_finding::Options::default_for_geometry(geometry),
+            border: border_finding::Options::default_for_geometry(geometry),
+        }
+    }
+}
+
+/// Find the timing mark grid in a ballot image.
+///
+/// # Errors
+///
+/// Returns an error if the timing mark grid cannot be found.
+#[allow(clippy::result_large_err)]
+pub fn find_timing_mark_grid(
+    ballot_image: &BallotImage,
+    geometry: &Geometry,
+    options: &Options,
+) -> Result<TimingMarks, Error> {
+    let shapes = shape_finding::BallotGridBorderShapes::from_ballot_image(
+        ballot_image,
+        geometry,
+        &options.shape,
+    );
+
+    ballot_image.debug().write("01-shapes", |canvas| {
+        shapes.debug_draw(canvas);
+    });
+
+    let candidates =
+        mark_finding::BallotGridCandidateMarks::from_shapes(ballot_image, geometry, shapes);
+
+    ballot_image.debug().write("02-candidate_marks", |canvas| {
+        candidates.debug_draw(canvas);
+    });
+
+    let corners = corner_finding::BallotGridCorners::find_all(
+        ballot_image.dimensions().into(),
+        geometry,
+        &candidates,
+        &options.corner,
+    )?;
+
+    ballot_image.debug().write("03-corners", |canvas| {
+        corners.debug_draw(canvas);
+    });
+
+    let borders = border_finding::BallotGridBorders::find_all(
+        geometry,
+        &corners,
+        &candidates,
+        &options.border,
+    )?;
+
+    ballot_image.debug().write("04-borders", |canvas| {
+        borders.debug_draw(canvas);
+    });
+
+    let [top_left_mark, top_right_mark, bottom_left_mark, bottom_right_mark] = [
+        corners.top_left(),
+        corners.top_right(),
+        corners.bottom_left(),
+        corners.bottom_right(),
+    ]
+    .map_cornerwise(|corner| corner.best_corner_grouping().corner_mark());
+
+    let [top_left_corner, top_right_corner, bottom_left_corner, bottom_right_corner] = [
+        top_left_mark,
+        top_right_mark,
+        bottom_left_mark,
+        bottom_right_mark,
+    ]
+    .map_cornerwise(|mark| mark.rect().center());
+
+    let timing_marks = TimingMarks {
+        geometry: geometry.clone(),
+        top_left_corner,
+        top_right_corner,
+        bottom_left_corner,
+        bottom_right_corner,
+        top_marks: borders.top.into_marks(),
+        bottom_marks: borders.bottom.into_marks(),
+        left_marks: borders.left.into_marks(),
+        right_marks: borders.right.into_marks(),
+        top_left_mark: *top_left_mark,
+        top_right_mark: *top_right_mark,
+        bottom_left_mark: *bottom_left_mark,
+        bottom_right_mark: *bottom_right_mark,
+    };
+
+    Ok(timing_marks)
+}
 
 #[derive(Debug, Clone, Serialize)]
 #[serde(rename_all = "camelCase")]
